+++ apps/koku-ui-ros/src/fed-modules/fleetHeatmapWrapper.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { FleetHeatmap } from 'routes/optimizations/fleetHeatmap';
+
+import { OptimizationsWrapper } from './optimizationsWrapper';
+
+const FleetHeatmapWrapper: React.FC = () => {
+  return (
+    <OptimizationsWrapper>
+      <FleetHeatmap />
+    </OptimizationsWrapper>
+  );
+};
+
+export default FleetHeatmapWrapper;